+++ djeddit/tests/test_views.py
@@ -1,5 +1,6 @@
 from django import VERSION as DJANGO_VERSION
 from django.test import TestCase
+
 if DJANGO_VERSION[:2] < (1, 10):
     from django.core.urlresolvers import reverse
 else:
@@ -19,8 +20,8 @@ class CreateThreadTest(TestCase, TestCalls):
     @classmethod
     def setUpTestData(cls):
         cls._setup_user(username='user', email='user@example.com', password='pass')
-        cls.topic = Topic.objects.create(title='Test_Topic')
-        cls.url = reverse('createThread', args=[cls.topic.title])
+        cls.topic = Topic.objects.create(title='Test Topic')
+        cls.url = reverse('createThread', args=[cls.topic.getUrlTitle()])
 
     def testLoads(self):
         self._test_call_view_loads(self.url, {})
@@ -40,6 +41,36 @@ class CreateThreadTest(TestCase, TestCalls):
         self._testSubmit(data)
 
 
+class DeleteTopicTest(TestCase, TestCalls):
+    def __init__(self, *args, **kwargs):
+        TestCase.__init__(self, *args, **kwargs)
+        TestCalls.__init__(self)
+
+    @classmethod
+    def setUpTestData(cls):
+        cls._setup_user(username='admin', email='admin@example.com', password='pass', is_superuser=True)
+        cls.topic = Topic.objects.create(title='Test Topic')
+        cls.url = reverse('deleteTopic', args=[cls.topic.getUrlTitle()])
+
+    def testDeleteTopic(self):
+        self.login()
+        self._test_call_view_submit(self.url, code=302)
+        self.assertRaises(Topic.DoesNotExist, self.topic.refresh_from_db)
+
+    def testUnknownTopic(self):
+        self.login()
+        url = reverse('deleteTopic', args=['Fake_Topic'])
+        self._test_call_view_code(url, 404)
+
+    def testRequireLogin(self):
+        self._test_call_view_redirected_login(self.url)
+
+    def testRequireSuperUser(self):
+        self._create_user_and_login()
+        self.login('user', 'pass')
+        self._test_call_view_redirected_login(self.url)
+
+
 class LockThreadTest(TestCase, TestCalls):
     def __init__(self, *args, **kwargs):
         TestCase.__init__(self, *args, **kwargs)
@@ -48,7 +79,7 @@ class LockThreadTest(TestCase, TestCalls):
     @classmethod
     def setUpTestData(cls):
         cls._setup_user(username='admin', email='admin@example.com', password='pass', is_superuser=True)
-        topic = Topic.objects.create(title='Test_Topic')
+        topic = Topic.objects.create(title='Test Topic')
         cls.thread = Thread.objects.create(title='Test_Thread', topic=topic, op=Post.objects.create())
         cls.url = reverse('lockThread', args=[cls.thread.id])
 
@@ -75,8 +106,9 @@ class LockThreadTest(TestCase, TestCalls):
         self.login()
         self._test_call_view_redirected_login(self.url)
 
-    def testUnauthenticated(self):
+    def testRequireLogin(self):
         self._test_call_view_redirected_login(self.url)
+
 
 class TopicsPageTest(TestCase, TestCalls):
     def __init__(self, *args, **kwargs):
@@ -95,16 +127,36 @@ class TopicPageTest(TestCase, TestCalls):
 
     @classmethod
     def setUpTestData(cls):
-        cls.topic = Topic.objects.create(title='Test_Topic')
+        cls._setup_user(username='admin', email='admin@example.com', password='pass', is_superuser=True)
+        cls.topic = Topic.objects.create(title='Test Topic')
+        cls.url = reverse('topicPage', args=[cls.topic.getUrlTitle()])
 
     def testLoads(self):
-        url = reverse('topicPage', args=[self.topic.getUrlTitle()])
-        self._test_call_view_loads(url)
+        self._test_call_view_loads(self.url)
 
-    def testRedirects(self):
+    def testUnknownTopic(self):
         url = reverse('topicPage', args=['Fake_Topic'])
-        redirected_url = reverse('topics')
-        self._test_call_view_redirects(url, redirected_url)
+        self._test_call_view_code(url, 404)
+
+    def testEditTopic(self):
+        self.login()
+        data = dict(title='Test Topic edited', description='Test Description')
+        self._test_call_view_submit(self.url, code=302, data=data)
+        self.topic.refresh_from_db()
+        self.assertEqual(self.topic.title, data['title'])
+        self.assertEqual(self.topic.description, data['description'])
+
+    def testEditTopicRequireLogin(self):
+        data = dict(title='Test Topic (edited)', description='Test Description')
+        self._test_call_view_submit(self.url, code=403, data=data)
+        self.topic.refresh_from_db()
+        self.assertNotEqual(self.topic.title, data['title'])
+        self.assertNotEqual(self.topic.description, data['description'])
+
+    def testEditTopicRequireSuperUser(self):
+        self._create_user_and_login()
+        self.login('user', 'pass')
+        self.testEditTopicRequireLogin()
 
 
 class ThreadPageTest(TestCase, TestCalls):
@@ -114,11 +166,11 @@ class ThreadPageTest(TestCase, TestCalls):
 
     @classmethod
     def setUpTestData(cls):
-        cls.topic = Topic.objects.create(title='Test_Topic')
+        cls.topic = Topic.objects.create(title='Test Topic')
         cls.thread = Thread.objects.create(title='Test_Thread', topic=cls.topic, op=Post.objects.create())
 
     def testLoads(self):
-        url = reverse('threadPage', args=[self.topic.title, self.thread.id])
+        url = reverse('threadPage', args=[self.topic.getUrlTitle(), self.thread.id])
         self._test_call_view_loads(url)
 
     def testWrongTopic(self):
@@ -127,7 +179,7 @@ class ThreadPageTest(TestCase, TestCalls):
         self._test_call_view_redirects(url, redirected_url)
 
     def testWrongThread(self):
-        url = reverse('threadPage', args=[self.topic.title, self.thread.id + 1])
+        url = reverse('threadPage', args=[self.topic.getUrlTitle(), self.thread.id + 1])
         redirected_url = reverse('topics')
         self._test_call_view_redirects(url, redirected_url)
 
@@ -140,7 +192,7 @@ class ReplyPostTest(TestCase, TestCalls):
     @classmethod
     def setUpTestData(cls):
         cls.post = Post.objects.create()
-        topic = Topic.objects.create(title='Test_Topic')
+        topic = Topic.objects.create(title='Test Topic')
         Thread.objects.create(title='Test_Thread', topic=topic, op=cls.post)
         cls.url = reverse('replyPost', args=[cls.post.uid])
 
@@ -149,7 +201,7 @@ class ReplyPostTest(TestCase, TestCalls):
 
     def testSubmit(self):
         self._test_call_view_submit(self.url, code=302, data=dict(content='replied'))
-        Post.objects.get(parent=self.post, content='replied') # assert object exists
+        Post.objects.get(parent=self.post, content='replied')  # assert object exists
 
     def testUnknownUid(self):
         uid = gen_uuid()
@@ -171,7 +223,7 @@ class EditPostTest(TestCase, TestCalls):
     def setUpTestData(cls):
         cls._setup_user(username='user', email='user@example.com', password='pass')
         cls.post = Post.objects.create(created_by=cls.user)
-        topic = Topic.objects.create(title='Test_Topic')
+        topic = Topic.objects.create(title='Test Topic')
         Thread.objects.create(title='Test_Thread', topic=topic, op=cls.post)
         user1 = createUser('user1', 'user1@example.com', 'pass')
         cls.User1Post = Post.objects.create(created_by=user1, parent=cls.post)
@@ -205,14 +257,14 @@ class EditPostTest(TestCase, TestCalls):
         self._test_call_view_code(url, 403)
 
     def testEditAnotherPostAsSuperuser(self):
-        self._login_as_admin()
+        self._create_user_and_login(create_admin=True)
         url = reverse('editPost', args=[str(self.User1Post.uid)])
         self._test_call_view_submit(url, code=302, data={'post-content': 'edited'})
         self.User1Post.refresh_from_db()
         self.assertEqual(self.User1Post.content, 'edited')
 
     def testEditThread(self):
-        self._login_as_admin()
+        self._create_user_and_login(create_admin=True)
         self._test_call_view_submit(self.url, code=302, data={'post-content': 'edited'})
         self.post.refresh_from_db()
         self.assertEqual(self.post.content, 'edited')
@@ -241,13 +293,17 @@ class VotePostTest(TestCase, TestCalls):
     def testSubmit(self):
         self.login()
         self.assertEqual(self.post.score, 0)
-        self._testVote(1, 1, 1) # test upvote
-        self._testVote(2, 1, 1) # test vote value above 1
-        self._testVote(-1, -1, -1) # test downvote
-        self._testVote(-2, -1, -1) # test downvote value below -1
+        self._testVote(1, 1, 1)  # test upvote
+        self._testVote(2, 1, 1)  # test vote value above 1
+        self._testVote(-1, -1, -1)  # test downvote
+        self._testVote(-2, -1, -1)  # test downvote value below -1
 
     def testRequiresLogin(self):
         self._test_call_view_redirected_login(self.url, dict(post=self.post.uid, vote=1))
+
+    def testEmptyData(self):
+        self.login()
+        self._test_call_view_code(self.url, 400, post=True)
 
     def testVoteOwnPost(self):
         self.login()
@@ -255,7 +311,7 @@ class VotePostTest(TestCase, TestCalls):
         self._test_call_view_submit(self.url, code=403, data=dict(post=post.uid, vote=1))
 
     def testVoteOwnPostAsSuperuser(self):
-        admin = self._login_as_admin()
+        admin = self._create_user_and_login(create_admin=True)
         post = Post.objects.create(created_by=admin)
         self._testVote(1, 1, 1, user=admin, post=post)
 
@@ -268,7 +324,7 @@ class DeletePostTest(TestCase, TestCalls):
     @classmethod
     def setUpTestData(cls):
         cls._setup_user(username='admin', email='admin@example.com', password='pass', is_superuser=True)
-        topic = Topic.objects.create(title='Test_Topic')
+        topic = Topic.objects.create(title='Test Topic')
         cls.thread = Thread.objects.create(title='Test_Thread', topic=topic, op=Post.objects.create())
         cls.comment = Post.objects.create(parent=cls.thread.op)
         cls.url = reverse('deletePost', args=[cls.comment.uid])
@@ -277,20 +333,19 @@ class DeletePostTest(TestCase, TestCalls):
         self._test_call_view_redirected_login(self.url)
 
     def testRequireSuperUser(self):
-        createUser(username='user', email='user@example.com', password='pass')
-        self.login(username='user', password='pass')
+        self._create_user_and_login()
         self._test_call_view_redirected_login(self.url)
 
     def testDeleteComment(self):
         self.login()
-        redirect_url = reverse('threadPage', args=[self.thread.topic.title, self.thread.id])
+        redirect_url = reverse('threadPage', args=[self.thread.topic.getUrlTitle(), self.thread.id])
         self._test_call_view_redirects(self.url, redirect_url)
         self.assertRaises(Post.DoesNotExist, self.comment.refresh_from_db)
 
     def testDeleteThread(self):
         self.login()
         url = reverse('deletePost', args=[self.thread.op.uid])
-        redirect_url = reverse('topicPage', args=[self.thread.topic.title])
+        redirect_url = reverse('topicPage', args=[self.thread.topic.getUrlTitle()])
         self._test_call_view_redirects(url, redirect_url)
         self.assertRaises(Post.DoesNotExist, self.thread.op.refresh_from_db)
         self.assertRaises(Thread.DoesNotExist, self.thread.refresh_from_db)
@@ -304,7 +359,7 @@ class UserSummaryTest(TestCase, TestCalls):
     @classmethod
     def setUpTestData(cls):
         cls._setup_user(username='user', email='user@example.com', password='pass')
-        cls.topic = Topic.objects.create(title='Test_Topic')
+        cls.topic = Topic.objects.create(title='Test Topic')
         cls.url = reverse('userSummary', args=[cls.username])
 
     def testLoadsEmpty(self):
@@ -313,7 +368,7 @@ class UserSummaryTest(TestCase, TestCalls):
     def testLoads(self):
         op = Post.objects.create(created_by=self.user)
         Thread.objects.create(topic=self.topic, title='Test Thread', op=op)
-        Post.objects.create(created_by=self.user)
+        Post.objects.create(created_by=self.user, parent=op)
         self._test_call_view_loads(self.url)
 
     def testInvalidUser(self):
@@ -329,7 +384,7 @@ class UserThreadsPageTest(TestCase, TestCalls):
     @classmethod
     def setUpTestData(cls):
         cls._setup_user(username='user', email='user@example.com', password='pass')
-        cls.topic = Topic.objects.create(title='Test_Topic')
+        cls.topic = Topic.objects.create(title='Test Topic')
         cls.url = reverse('userThreads', args=[cls.username])
 
     def testLoadsEmpty(self):
@@ -353,7 +408,7 @@ class UserRepliesPageTest(TestCase, TestCalls):
     @classmethod
     def setUpTestData(cls):
         cls._setup_user(username='user', email='user@example.com', password='pass')
-        topic = Topic.objects.create(title='Test_Topic')
+        topic = Topic.objects.create(title='Test Topic')
         cls.thread = Thread.objects.create(topic=topic, title='Test Thread', op=Post.objects.create())
         cls.url = reverse('userThreads', args=[cls.username])
 
@@ -367,3 +422,70 @@ class UserRepliesPageTest(TestCase, TestCalls):
     def testInvalidUser(self):
         url = reverse('userSummary', args=['invaliduser'])
         self._test_call_view_code(url, 404)
+
+
+class UsersPageTest(TestCase, TestCalls):
+    def __init__(self, *args, **kwargs):
+        TestCase.__init__(self, *args, **kwargs)
+        TestCalls.__init__(self, 'djeddit/users_page.html')
+
+    @classmethod
+    def setUpTestData(cls):
+        cls._setup_user(username='admin', email='admin@example.com', password='pass', is_superuser=True)
+        cls.url = reverse('usersPage')
+
+    def testLoads(self):
+        user = createUser()
+        topic = Topic.objects.create(title='Test Topic')
+        op = Post.objects.create(created_by=user)
+        Thread.objects.create(topic=topic, title='Test Thread', op=op)
+        Post.objects.create(created_by=user, parent=op)
+        self.login()
+        self._test_call_view_loads(self.url)
+
+    def testRequireLogin(self):
+        self._test_call_view_redirected_login(self.url)
+
+    def testRequireSuperUser(self):
+        self._create_user_and_login()
+        self._test_call_view_redirected_login(self.url)
+
+
+class SetUserStatusTest(TestCase, TestCalls):
+    def __init__(self, *args, **kwargs):
+        TestCase.__init__(self, *args, **kwargs)
+        TestCalls.__init__(self)
+
+    @classmethod
+    def setUpTestData(cls):
+        cls._setup_user(username='admin', email='admin@example.com', password='pass', is_superuser=True)
+        cls.url = reverse('setUserStatus')
+
+    def _testChangeStatus(self, user, status):
+        self._test_call_view_submit(self.url, data=dict(username=user.username, status=status))
+        user.refresh_from_db()
+        if status == 'active':
+            self.assertTrue(user.is_active)
+            self.assertTrue(not user.is_superuser)
+        elif status == 'banned':
+            self.assertTrue(not user.is_active)
+            self.assertTrue(not user.is_superuser)
+        elif status == 'admin':
+            self.assertTrue(user.is_active)
+            self.assertTrue(user.is_superuser)
+        else:
+            raise Exception('Invalid status: %s' % status)
+
+    def testPromoteToAdmin(self):
+        user = self._create_user_and_login()
+        self.login()
+        self._testChangeStatus(user, 'admin')
+        self._testChangeStatus(user, 'active')
+        self._testChangeStatus(user, 'banned')
+
+    def testRequireLogin(self):
+        self._test_call_view_redirected_login(self.url)
+
+    def testRequireSuperUser(self):
+        self._create_user_and_login()
+        self._test_call_view_redirected_login(self.url)